requirements.txt
Flask
Pinecone
google-generativeai
numpy
tqdm
python-dotenv
